Changed detection errors to give you the name of the blocks that are causing the problem, rather than IDs. Also some additional progress on the repair system, but it isn't done yet.

+++ src/main/java/net/countercraft/movecraft/config/Settings.java
@@ -39,6 +39,6 @@ public class Settings {
 	public static int ManOverBoardTimeout = 60;
 	public static int FireballLifespan = 6;
 	public static int RepairTicksPerBlock = 0;
-	public static int RepairMoneyPerBlock = 0;
+	public static double RepairMoneyPerBlock = 0.0;
 	public static boolean FireballPenetration = true;
 }
